Simplify template
