+++ template_content/README.md.jinja
@@ -26,7 +26,7 @@ This project has been generated using AlgoKit. See below for default getting sta
    - Other
      1. Open the repository root in your text editor of choice
      2. In a terminal run `poetry shell`
-     3. Run `python -m smart_contracts` through your debugger of choice
+     3. Run `python app.py` through your debugger of choice
 
 ### Subsequently
