--- template_content/smart_contracts/.env.template.jinja
@@ -1,6 +1,0 @@
-ALGOD_TOKEN={{ algod_token }}
-ALGOD_SERVER={{ algod_server }}
-ALGOD_PORT={{ algod_port }}
-INDEXER_TOKEN={{ indexer_token }}
-INDEXER_SERVER={{ indexer_server }}
-INDEXER_PORT={{ indexer_port }}
