+++ template_content/{% if ide_vscode %}.vscode{% endif %}/settings.json.jinja
@@ -31,7 +31,6 @@
     }
   },
   "python.analysis.indexing": true,
-  "python.analysis.extraPaths": ["${workspaceFolder}/smart_contracts"],
   "python.formatting.provider": "black",
   "[python]": {
     // https://dev.to/eegli/quick-guide-to-python-formatting-in-vs-code-2040
